Initial skeleton for Rust-based rewrite (unofficial)
This is the initial skeleton for the `develop-rewrite-rust` branch
starting point that I - @shymega - am working on in my free time.

It's not official, but I've got it in a branch on the main repo to make
development easier.

I may move it at some point.

Consider this EXPERIMENTAL, and not at all something to be used yet.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>